+++ Henrique-Sudoku/Sudoku-GUI-otimizado/testes/relatorio_performance.txt
