+++ Henrique-Sudoku/Sudoku-GUI-otimizado/testes/tempo_execucao_place.png
