+++ Henrique-Sudoku/Sudoku-GUI-otimizado/testes/tempo_execucao_solve.png
